requirements.txt
django==2.2.11
djangorestframework==3.11.0
psycopg2-binary==2.8.4
Pillow>=5.3.0,<5.4.0